+++ sentry-android-fragment/src/test/java/io/sentry/android/fragment/FragmentLifecycleIntegrationTest.kt
@@ -12,7 +12,6 @@ import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
 import io.sentry.IHub
 import io.sentry.SentryOptions
-import io.sentry.android.fragment.internal.SentryFragmentLifecycleCallbacks
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
@@ -92,7 +91,7 @@ class FragmentLifecycleIntegrationTest {
             check { fragmentCallbacks ->
                 val callback = (fragmentCallbacks as SentryFragmentLifecycleCallbacks)
                 assertTrue(callback.enableAutoFragmentLifecycleTracing)
-                assertEquals(emptySet(), callback.loggedFragmentLifecycleStates)
+                assertEquals(emptySet(), callback.filterFragmentLifecycleBreadcrumbs)
             },
             eq(true)
         )